requirements.txt
dnspython
geoip2